# InvServer.Api/wwwroot/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InvServer API",
    "description": "Comprehensive API for Inventory Management, Workflow Governance, and Idempotent Fulfillment.\n\nKey Features:\n- **Idempotency**: All fulfillment operations require `X-Idempotency-Key`.\n- **Workflow**: Governed lifecycle (Draft -> In Workflow -> Fulfillment).\n- **RBAC**: Scope-based access control (OWN, DEPT, WAREHOUSE, GLOBAL).\n\n(Updated to reflect latest Admin/Workflow/Inventory controller changes.)",
    "version": "1.2.0",
    "contact": {
      "name": "API Support",
      "email": "support@invserver.local"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiResponse_Generic": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "correlationId": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
...
      "ApiResponse_Long": {
        "type": "object",
        "properties": {
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "correlationId": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        }
      },
...
      "SubmitRequestDto": {
        "type": "object",
        "properties": {
          "manualAssignments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/WorkflowManualAssignmentDto"
            }
          }
        }
      },
...
    },
    "parameters": {
      "X-Idempotency-Key": {
        "name": "X-Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Unique key to ensure idempotency. Example: `550e8400-e29b-41d4-a716-446655440000`"
      },
      "X-Correlation-Id": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Tracing ID for log correlation."
      }
    }
  },
...
  "paths": {
...
    "/api/inventory/requests": {
      "get": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "List Requests",
        "description": "Lists inventory requests visible to the user based on RBAC scope (OWN, DEPT, WAREHOUSE, GLOBAL).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "Create Draft Request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Long"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/requests/{requestId}": {
      "get": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "Get Request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "Update Draft Request",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      }
    },
    "/api/inventory/requests/{requestId}/submit": {
      "post": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "Submit Request",
        "description": "Transitions request from DRAFT to IN_WORKFLOW.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequestDto"
              }
            }
          }
        }
      }
    },
    "/api/inventory/requests/{requestId}/cancel": {
      "post": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "Cancel Request",
        "description": "Cancels request. Only allowed if NO active stock reservations exist.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled"
          }
        }
      }
    },
...
    "/api/inventory/requests/{requestId}/fulfillment/reserve": {
      "post": {
        "tags": [
          "Fulfillment"
        ],
        "summary": "Reserve Stock",
        "description": "Idempotent operation to reserve stock. Only allowed if workflow step is FULFILLMENT.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/X-Idempotency-Key"
          },
          {
            "$ref": "#/components/parameters/X-Correlation-Id"
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Required for idempotent fulfillment operations."
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional correlation id (defaults to trace identifier)."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockMovementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (or Replay)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Long"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (Missing Key or Invalid State)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          },
          "409": {
            "description": "Key Conflict (Processing)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/requests/{requestId}/fulfillment/release": {
      "post": {
        "tags": [
          "Fulfillment"
        ],
        "summary": "Release Stock",
        "description": "Reverses a reservation.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "$ref": "#/components/parameters/X-Idempotency-Key"
          },
          {
            "$ref": "#/components/parameters/X-Correlation-Id"
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Required for idempotent fulfillment operations."
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional correlation id (defaults to trace identifier)."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockMovementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Long"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inventory/requests/{requestId}/history": {
      "get": {
        "tags": [
          "Inventory Requests"
        ],
        "summary": "Get Request History",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Generic"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}